+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 - Now catching errors in `!image` creation and letting the user know their request failed.
 - Make chat `!personality` clearer in the prompt.
 - Fixed a bug where `!!` and `! ` were being treated as commands. Now only `!` followed by at least 1 alphanumeric character is treated as a command. Thanks @dannovikov @fitztrev!
+- Type lichess response for `!bbb` command and limit tournaments returned to prevent lag in first use.
 
 ### 1.10.1 (2024-02-05)